chore: rename `case-table-card-common` to `case-tile-common`, etc.
