+++ v3/src/components/case-card/case-card.v2.js
@@ -6,8 +6,8 @@ import { preventCollectionReorg } from "../../utilities/plugin-utils"
 import { createReactFactory, DG } from "../../v2/dg-compat.v2"
 import { SC } from "../../v2/sc-compat"
 import { getSharedCaseMetadataFromDataset } from "../../models/shared/shared-data-utils"
-import { EditAttributePropertiesModal } from "../case-table-card-common/attribute-menu/edit-attribute-properties-modal"
-import { EditFormulaModal } from "../case-table-card-common/attribute-menu/edit-formula-modal"
+import { EditAttributePropertiesModal } from "../case-tile-common/attribute-menu/edit-attribute-properties-modal"
+import { EditFormulaModal } from "../case-tile-common/attribute-menu/edit-formula-modal"
 
 import "./attribute-name-cell.v2"
 import "./attribute-value-cell.v2"
