+++ v3/src/components/case-card/case-view.tsx
@@ -5,7 +5,7 @@ import { ICollectionModel } from "../../models/data/collection"
 import { CaseAttrsView } from "./case-attrs-view"
 import { useCaseCardModel } from "./use-case-card-model"
 import { CollectionContext, ParentCollectionContext, useCollectionContext } from "../../hooks/use-collection-context"
-import { CollectionTitle } from "../case-table-card-common/collection-title"
+import { CollectionTitle } from "../case-tile-common/collection-title"
 import { t } from "../../utilities/translation/translate"
 import { IAttribute } from "../../models/data/attribute"
 import { createAttributesNotification, createCasesNotification } from "../../models/data/data-set-notifications"
