+++ v3/src/components/case-table/case-table-inspector.tsx
@@ -7,7 +7,7 @@ import TrashIcon from "../../assets/icons/icon-trash.svg"
 import ValuesIcon from "../../assets/icons/icon-values.svg"
 import { DatasetInfoModal } from "./inspector-panel/dataset-info-modal"
 import { TrashMenuList } from "./inspector-panel/trash-menu-list"
-import { HideShowMenuList } from "../case-table-card-common/hide-show-menu-list"
+import { HideShowMenuList } from "../case-tile-common/hide-show-menu-list"
 import { t } from "../../utilities/translation/translate"
 import { RulerMenuList } from "./inspector-panel/ruler-menu-list"
 import { ITileInspectorPanelProps } from "../tiles/tile-base-props"
