+++ v3/src/components/case-table/column-header.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useRef, useState } from "react"
-import { AttributeHeader } from "../case-table-card-common/attribute-header"
+import { AttributeHeader } from "../case-tile-common/attribute-header"
 import { TRenderHeaderCellProps } from "./case-table-types"
 import { ColumnHeaderDivider } from "./column-header-divider"
