+++ v3/src/components/case-table/use-index-column.tsx
@@ -5,7 +5,7 @@ import { createPortal } from "react-dom"
 import { kIndexColumnKey, kInputRowKey, TColSpanArgs, TColumn, TRenderCellProps } from "./case-table-types"
 import { ColumnHeader } from "./column-header"
 import { IndexMenuList } from "./index-menu-list"
-import { useParentChildFocusRedirect } from "../case-table-card-common/use-parent-child-focus-redirect"
+import { useParentChildFocusRedirect } from "../case-tile-common/use-parent-child-focus-redirect"
 import { useCaseMetadata } from "../../hooks/use-case-metadata"
 import { useCollectionContext } from "../../hooks/use-collection-context"
 import { useDataSetContext } from "../../hooks/use-data-set-context"
