+++ v3/src/components/case-tile-common/attribute-menu/attribute-menu.scss
