+++ v3/src/components/case-tile-common/attribute-menu/edit-attribute-properties-modal.tsx
