+++ v3/src/components/case-tile-common/case-tile-title-bar.scss
