+++ v3/src/components/case-tile-common/case-tile-types.ts
